@@ -1,4 +1,4 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
 	name: "fxSDEngine",
 	platforms: [
-		.iOS(.v17),
+		.iOS(.v18),
 	],
 	products: [
 		// Products define the executables and libraries a package produces, making them visible to other packages.
