@@ -6,7 +6,7 @@ import fXDKit
 
 
 @Observable
-open class fxSDengineBasic: @preconcurrency SDEngine {
+open class fxSDengineBasic: @preconcurrency SDEngine, @unchecked Sendable {
     public var mainSDNetworking: any SDNetworking
 	required public init(mainSDNetworking: SDNetworking) {
         self.mainSDNetworking = mainSDNetworking
@@ -55,11 +55,13 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
     public var lastHTTPURLResponses: [HTTPURLResponse] = []
 
 	open func action_Synchronize() {
-        Task {	@MainActor in
+        Task {
             let error = try await synchronize_withSystem()
             let _ = await refresh_allModels()
 
-            UIAlertController.errorAlert(error: error, title: "Possibly, your Stable Diffusion server is not operating.")
+            DispatchQueue.main.async {
+                UIAlertController.errorAlert(error: error, title: "Possibly, your Stable Diffusion server is not operating.")
+            }
         }
 	}
 
@@ -86,11 +88,9 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
 
         let loadedPayload = try SDcodablePayload.loaded(from: fileURL.jsonURL, withControlNet: true)
 
-        await MainActor.run {
-            nextPayload = loadedPayload
+        nextPayload = loadedPayload
 
-            selectedImageURL = fileURL
-        }
+        selectedImageURL = fileURL
 
         return error_2
     }
@@ -108,9 +108,7 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
             fxdPrint(name: "INTERNAL_SYSINFO", dictionary: jsonDictionary)
         }
 #endif
-        await MainActor.run {
-            systemInfo = data?.decode(SDcodableSysInfo.self)
-        }
+        systemInfo = data?.decode(SDcodableSysInfo.self)
 
         return error
     }
@@ -124,34 +122,42 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
     }
 
 	open func action_ChangeCheckpoint(_ checkpoint: SDcodableCheckpoint) {
-        Task {	@MainActor in
+        Task {
             let error_0 = await change_systemCheckpoints(checkpoint: checkpoint)
 
 			guard error_0 == nil else {
-                UIAlertController.errorAlert(error: error_0)
+                DispatchQueue.main.async {
+                    UIAlertController.errorAlert(error: error_0)
+                }
 				return
 			}
 
 
             let error_1 = await refresh_systemInfo()
 
-            UIAlertController.errorAlert(error: error_1)
+            DispatchQueue.main.async {
+                UIAlertController.errorAlert(error: error_1)
+            }
         }
 	}
 
     open func action_ChangeVAE(_ vae: SDcodableVAE) {
-        Task {    @MainActor in
+        Task {
             let error_0 = await change_systemVAE(vae: vae)
 
             guard error_0 == nil else {
-                UIAlertController.errorAlert(error: error_0)
+                DispatchQueue.main.async {
+                    UIAlertController.errorAlert(error: error_0)
+                }
                 return
             }
 
 
             let error_1 = await refresh_systemInfo()
 
-            UIAlertController.errorAlert(error: error_1)
+            DispatchQueue.main.async {
+                UIAlertController.errorAlert(error: error_1)
+            }
         }
     }
 
@@ -250,22 +256,20 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
             fxdPrint("\(String(describing: T.self))", (jsonObject as? Array<Any>)?.count)
         }
 #endif
-        await MainActor.run {
-            let models = data?.decode(Array<T>.self) ?? []
+        let models = data?.decode(Array<T>.self) ?? []
 
-            switch T.self {
-                case is SDcodableCheckpoint.Type:
-                    systemCheckpoints = models as? Array<SDcodableCheckpoint> ?? []
-                case is SDcodableVAE.Type:
-                    systemVAEs = SDcodableVAE.defaultArray() + (models as? Array<SDcodableVAE> ?? [])
-                case is SDcodableSampler.Type:
-                    systemSamplers = models as? Array<SDcodableSampler> ?? []
-                case is SDcodableScheduler.Type:
-                    systemSchedulers = models as? Array<SDcodableScheduler> ?? []
+        switch T.self {
+            case is SDcodableCheckpoint.Type:
+                systemCheckpoints = models as? Array<SDcodableCheckpoint> ?? []
+            case is SDcodableVAE.Type:
+                systemVAEs = SDcodableVAE.defaultArray() + (models as? Array<SDcodableVAE> ?? [])
+            case is SDcodableSampler.Type:
+                systemSamplers = models as? Array<SDcodableSampler> ?? []
+            case is SDcodableScheduler.Type:
+                systemSchedulers = models as? Array<SDcodableScheduler> ?? []
 
-                default:
-                    break
-            }
+            default:
+                break
         }
 
         return error
@@ -511,9 +515,7 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
 
 
         guard !didInterrupt else {
-            await MainActor.run {
-                didInterrupt = false
-            }
+            didInterrupt = false
 
             let interruptedError = SDError(
                 domain: "SDEngine",
@@ -537,12 +539,10 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
             utilizedControlNet: utilizedControlNet)
 
 
-        await MainActor.run {
             nextPayload = newPayload
             nextPayload?.userConfiguration?.controlnet = utilizedControlNet
 
             selectedImageURL = newImageURL
-        }
 
         return error
     }
@@ -586,7 +586,7 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
 
 
     open func continueMonitoring() {
-        Task {	@MainActor in
+        Task {
             let (newProgress, isSystemBusy, _) = await monitor_progress(quiet: true)
             if newProgress != nil || (didStartGenerating || isSystemBusy) != self.isSystemBusy {
                 monitoredProgress = newProgress
@@ -645,7 +645,9 @@ open class fxSDengineBasic: @preconcurrency SDEngine {
                 NSLocalizedFailureReasonErrorKey: "Server's image generating is canceled",
             ])
 
-        UIAlertController.errorAlert(error: interruptedError)
+        DispatchQueue.main.async {
+            UIAlertController.errorAlert(error: interruptedError)
+        }
 
         return interruptedError
     }
